The game doesn't crash when you try to buy an item. Doesn't work though.

+++ src/shopPackage/Shop.java
@@ -12,7 +12,7 @@ public class Shop {
 	private ShopKeeper keeper;
 	private static int shopCount;
 	private int maxSize;
-	
+
 	public Shop() {
 		this.maxSize = this.determineMax();
 		this.itemList = new Item[maxSize];
@@ -20,65 +20,70 @@ public class Shop {
 		Shop.shopCount++;
 		this.createShopInventory();
 	}
-	
+
 	private int determineMax() {
 		int max = 1;
 		int mod = Battlefield.getBattleCount();
 		if (mod < 100) {
 			max = (mod / 5) + 1;
+		} else {
+			max = 20;
 		}
-		else {
-			max = 20; }
 		return max;
 	}
 
 	private void createShopInventory() {
 		for (int i = 0; i < itemList.length; i++) {
-			this.itemList[i] = new Item(this.insertWeightedItem(),this.keeper.getPerson());
+			this.itemList[i] = new Item(this.insertWeightedItem(), this.keeper.getPerson());
 		}
 	}
 
 	private ItemNames insertWeightedItem() {
 		int mod = this.calculateItemWeight();
 		ItemNames itemN = ItemNames.InconceivableItem;
-		while(itemN.getWeight() > mod) {
+		while (itemN.getWeight() > mod) {
 			itemN = ItemNames.getRandom();
-			}
-		return itemN;
 		}
+		return itemN;
+	}
 
 	private int calculateItemWeight() {
-		int itemWeight = (Battlefield.getBattleCount()/4)+1;
+		int itemWeight = (Battlefield.getBattleCount() / 4) + 1;
 		return itemWeight;
 	}
-	
+
+	// TODO Fix this
 	private void deleteItem(Integer index) {
 		this.itemList[index] = null;
 	}
 
 	private Integer findIndex(String name) {
 		Integer index = 1;
-		for (int i = 0; i < this.itemList.length; i++) {
+		for (int i = 0; i < this.itemList.length;) {
 			if (this.itemList[i].equals(name)) {
 				index = i;
+				return index;
 			} else {
-				System.out.println("Item not found.");
 				index = null;
+				return index;
 			}
 		}
-		return index;
+		return null;
 	}
-	//Everything breaks here. Throws a NullPointerException
-	//TODO Fix this.
+
 	public void buyItem(Character player, String choice) {
-				Integer index = this.findIndex(choice);
-				this.deleteItem(index);
-				player.getInventory().addItem(this.itemList[index].getItemName());
-				System.out.println("Enter the name of the Item you wish to buy. Type 0 to leave this menu.");
+		Integer index = this.findIndex(choice);
+		if (index != null) {
+			player.getInventory().addItem(this.itemList[index].getItemName());
+			System.out.println(player.getName() + " buys " + this.itemList[index].getName());
+			this.deleteItem(index);
+		} else {
+			System.out.println("item not found.");
+		}
 	}
-	
+
 	public String toString() {
-		String returnString = "Items in the Shop of "+this.keeper.getName()+ " : \n";
+		String returnString = "Items in the Shop of " + this.keeper.getName() + " : \n";
 		for (int i = 0; i < itemList.length; i++) {
 			if (itemList[i] == null)
 				returnString += "";
@@ -88,5 +93,5 @@ public class Shop {
 		return returnString;
 
 	}
-	
+
 }
